requirements.txt
anyio==2.1.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
biopython==1.78
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cutadapt==3.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dnaio==0.5.0
entrypoints==0.3
idna==3.1
importlib-metadata==3.5.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-resource-usage==0.5.1
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-system-monitor==0.8.0
jupyterlab-topbar==0.6.1
jupytext==1.10.2
kiwisolver==1.3.1
markdown-it-py==0.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mdit-py-plugins==0.2.5
mistune==0.8.4
multiprocess==0.70.11.1
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathos==0.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
xopen==1.0.1
zipp==3.4.0
